+++ index.html
@@ -7,6 +7,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
+  <script defer type="module" src="js/main.js"></script>
 </head>
 
 <body>
